datapath: Inner And Outer Flows
Allow a flow to be visible in two forms, an outer flow and an inner flow.
This may occur if the actions allow further decoding of a packet to
provide a more fine-grained match. In this case the first-pass
coarse-grained match will hit the outer-flow and the second-pass
fined-grained match will hit the inner-flow.

Inner-flows are not visible to user-space. Rather, they are just an
internal representation to handle the case of actions allowing further
packet decoding. An inner-flow is associated with its outer-flow and
deleted when the outer-flow is deleted. An outer-flow may have more than
one inner-flow but an inner-flow may only have one outer-flow.

For example:

In the case of MPLS, L3 and L4 information may not initially be decoded
from the frame as the ethernet type of the frame is an MPLS type and no
information is known about the type of the inner frame.

However, the type of the inner frame may be provided by an mpls_pop action
in which case L3 and L4 information may be decoded providing a finer
grained match than is otherwise possible.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

v2.20 (draft)
* Use encap() for MPLS-encapsulated ODP flow key data

v2.19
* Update ovs_actions_allow_l3_extraction() to always return
  htons(0) instead of 0.
* enhance ovs_actions_allow_l3_extraction to look for the first
  MPLS POP action with a non-MPLS ethertype rather than the first
  MPLS POP action.

* Correct logic error in ovs_flow_tbl_next() that prevented facets from
  being dumped by the kernel-datapath. The result being that ovs-vswitchd
  would evict facts from the datapath regardless if they had been
  recently used or not. This may in turn cause flow misses and upcalls
  to ovs-vswitchd.

  This change should be squashed into "datapath: Inner And Outer Flows"
  I plan to do so for the v2.19 post of the MPLS patch-set.

v2.18
* No change

v2.17
* Rebase
* Do not allocate and free actions for outer flows,
  the actions of an inner flow may be shared.

v2.16
* No change

v2.15
* Rebase

v2.14
* No change

v2.13
* Merge "datapath: Allow inner_flows" and
  "datapath: re-lookup a flow if actions allow a more fine grained match"
  and rename the resulting patch "datapath: Inner and Outer Flows".
* As suggested by Jarno Rajahalme
  - Add encap_eth_type to struct sw_flow and use it to store the
    ethernet type supplied by actions of a flow if any. This is
    an optimisation to avoid scanning the actions for each packet.

v2.12
* No change

v2.11
* First post