dpif-netdev: Inner And Outer Flows
This is the user-space datapath component of this change

Allow a flow to be visible in two forms, an outer flow and an inner flow.
This may occur if the actions allow further decoding of a packet to
provide a more fine-grained match. In this case the first-pass
coarse-grained match will hit the outer-flow and the second-pass
fined-grained match will hit the inner-flow.

Inner-flows are not visible to user-space. Rather, they are just an
internal representation to handle the case of actions allowing further
packet decoding. An inner-flow is associated with its outer-flow and
deleted when the outer-flow is deleted. An outer-flow may have more than
one inner-flow but an inner-flow may only have one outer-flow.

For example:

In the case of MPLS, L3 and L4 information may not initially be decoded
from the frame as the ethernet type of the frame is an MPLS type and no
information is known about the type of the inner frame.

However, the type of the inner frame may be provided by an mpls_pop action
in which case L3 and L4 information may be decoded providing a finer
grained match than is otherwise possible.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

v2.19
* Update dpif_netdev_actions_allow_l3_extraction() to so that
  it returns when it finds the first MPLS_POP action with a
  non-MPLS type rather than the first MPLS_POP action.

v2.18
* Update match_wc_init() to use dl_type_is_ip_any()

v2.14 - v2.17
* No change

v2.13
* First post

enhance dpif_netdev_actions_allow_l3_extraction