ofproto: Enhance checking of ethertype of mpls_pop actions
The current check is not alwaus correct. It ensures ensures that the ether
type of an mpls_pop action is never an MPLS ether type. If a packet has
more than LSE then popping the outer-most LSE will require an mpls_pop
action with an MPLS ether type.

Replace this with logic in execute_mpls_pop_action to processing on invalid
mpls_pop action and check the validity of the ether type based on the
MPLS stack depth.

Also add a test that exercises this by performing an mpls_pop
on a frame with two MPLS Label Stack Entries.

Signed-off-by: Simon Horman <horms@verge.net.au>