packet: Always update ethernet type on pop_mpls
It seems to me that regardless of if the bottom of the MPLS label stack is
reached or not, the ethtype of a pop_mpls actions should be used to update
the ethtype of the flow or packet.

A test has been added to exercise this logic.

Also Limit scope of mh in pop_mpls().

Signed-off-by: Simon Horman <horms@verge.net.au>