ofproto-dpif: Ensure BOS bit of innermost MPLS LSE is set on push
This resolves a situation where the BOS bit of the innermost MPLS LSE
will not be set if all LSEs are popped and then an LSE is pushed.

e.g.
Packet with a single MPLS LSE has the following actions applied:
	pop_mpls:0x800
	push_mpls:0x8847

This patch also adds a test for this condition to the test suite.

Signed-off-by: Simon Horman <horms@verge.net.au>