ofproto-dpif: Update mpls_lse and inner_mpls_lse on MPLS push and pop
Update the mpls_lse to inner_mpls_lse on mpls_pop.
And update the inner_mpls_lse on pop.

This extends the cases where the LSE is tracked the correctly to
a depth of two. For example, the following case is now handled correctly.

A packet with a non-empty MPLS label stack has the following actions applied.
	push_mpls:0x8847
	pop_mpls:0x0800
	push_mpls:0x8847

A test has been added to the test suite for the above case to exercise
this logic.

This does not work correctly if more than two mpls_pop actions are
applied reducing the stack depth by two and an mpls_push action is then
applied. However at this time there are restrictions in place
to disallow multiple mpls_pop actions.

Signed-off-by: Simon Horman <horms@verge.net.au>