ofp-util: Ensure MPLS LSE is correct after mpls_pop
There is insufficient information tracked to determine if after pushing the
outermost MPLS LSE has been modified or .  So if there is an outermost MPLS
LSE persent then always set it to the value in flow. This should be correct
but may result in an unnecessary set action in some cases.

For example:

The following actions are applied to a frame with at least two
MPLS LSE entries:
	pop_mpls:0x8847
	load:21->OXM_OF_MPLS_LABEL[]

Without this change the result will be incorrect unless the
label of the second outermost LSE of the original frame was 21.

A test had been also added to exercise this condition.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

For correctness this patch depends on
"ofproto-dpif: Update mpls_lse and inner_mpls_lse on MPLS push and pop"