flow: Exclude mpls_depth and encap_dl_type from flow match
mpls_depth and encap_dl_type are not required as part of the match
and in some cases can lead to a false negative.

For example, a packet with two MPLS label stack entries will be have a flow
with mpls_depth calculated by flow_extract() which uses eth_type_mpls() to
count the number of MPLS LSE present. However a key received from the
datapath for a facet may only have a single MPLS attribute present and
odp_flow_key_to_flow() will create a flow with an mpls_depth of only 1.

It would probably be nicer to move mpls_depth and encap_dl_type out
of struct flow, however this appears to require a rather invasive
refactoring in order to provide an alternate place to store mpls_depth and
encap_dl_type. So instead struct flow has been partitioned into portions
to match followed by portions not to match.

Signed-off-by: Simon Horman <horms@verge.net.au>