ofproto-dpif: Update facet_revalidate for inner_flows
If a subfacet has an inner_flow then the inner_flow should
be revalidated in inner_flows.

To allow this add an inner_flow element to subfacet.
Subfacets belong to their outer_flow and thus the
iteration over subfacets that occurs in facet_revalidate() will
never occur when facet is an inner facet.

Signed-off-by: Simon Horman <horms@verge.net.au>