ovs-ofctl: Handle any number of buckets in group statistics
struct ofputil_group_stats has an arbitrary limit
of 16 buckets for which it can record statistics.
However the code does not appear to enforce this
limit and it seems to me that the code could overflow.

This patch aims to remove the arbitrary limit by
removing the 'bucket_stats' field from struct ofputil_group_stats
and instead allowing an instance of that structure to be
followed by an any number of struct bucket_counter.

Signed-off-by: Simon Horman <horms@verge.net.au>

---
v2
* No change