ovs-ofctl: Handle any number of buckets in group statistics
struct ofputil_group_stats has an arbitrary limit
of 16 buckets for which it can record statistics.
However the code does not appear to enforce this
limit and it seems to me that the code could overflow.

This patch aims to remove the arbitrary limit by
changing the 'bucket_stats' field of struct ofputil_group_stats
from a fixed length array to a pointer whose storage is allocated and freed
as necessary.

Signed-off-by: Simon Horman <horms@verge.net.au>

---
v3
* As suggested by Ben Pfaff
  - Vastly simplify the change by using an explicit pointer for
    the 'bucket_stats' member of struct ofputil_group_stats rather
    than implicit variable length array appended to the end of the
    structure.

v2
* No change