Add support for write-actions
Implementation note:

All actions which modify a field are added to the action set
at the point where "set" actions should be added. In general
modifying a field many times is the same as only modifying it
the last time so the implementation simply adds all set actions to
the action set in the order they are specified. However, this breaks
down if two actions modify different portions of the same field.

Some examples.

1. load acting a subfield
2. mod_vlan_vid, mod_vlan_pcp

If this is considered to be a problem one possible solution would be to
either disallow all set actions other than set_field in write_actions.
Another possible solution is prohibit problematic the actions listed above
in write actions.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

v6
* No change

v5
* Add OFPACT_DEC_MPLS_TTL to ofpacts_list_to_action_set()
  This was left out of the OFPACT_DEC_MPLS_TTL update made
  in v4.

v4
* Rebase
* As suggested by Jarno Rajahalme
  - Group instructions in ofpact_is_allowed_in_actions_set()
    and add comment about why they are exclude from the actions set.
  - Allow OFPACT_DEC_MPLS_TTL in action set.
  - Add note about why some non-OpenFlow actions are excluded from
    the action set

v3
* No change

v2
* First post