Translation of indirect and all groups
Allow translation of indirect and all groups.
Also allow insertion of indirect and all groups by
changing the maximum permitted number in the groups
table from 0 to OFPG_MAX.

Implementation note:

After translating the actions for each bucket
ctx->base_flow is reset to its state prior to translation
of the buckets actions. This is equivalent to cloning the
bucket before applying actions. This is my interpretation
of the OpenFlow 1.3.2 specification section 5.6.1 Group Types,
which includes the following text. I believe there is room
for other interpretations.

* On all groups: "The packet is effectively cloned for each bucket; one
  packet is processed for each bucket of the group."
* On indirect groups: "This group type is effectively identical to an
  all group with one bucket."

Signed-off-by: Simon Horman <horms@verge.net.au>

---
v3
* Rebase for "ofproto-dpif: Hide struct rule_dpif internally"
* Update for hidden group_dpif
* Corrected typo in all groups test

v2
* First post