connmgr: Remove unnecessary reason fixup logic
A packet_in message may be sent for one of two reasons.

1. As the result of an controller action supplied in a rule.
   This is executed if a packet matches the match for the rule.
   The packet_in reason is explicitly part of the action and
   is thus correct.

2. As the result of the failure to match a packet against any rule.
   In this case a rule present in TBL_INTERNAL is used. This rule
   has the packet_in reason set to OFPR_NO_MATCH.

   This is the behaviour described in OpenFlow 1.1 and 1.2 when
   a miss occurs. And to date it is the behaviour provided
   by Open vSwtich for all OpenFlow versions.

   When this behaviour is in effect OFPR_NO_MATCH is the desired
   packet_in reason and by virtue of the TBL_INTERNAL rule described
   above is always that value.

So for both cases the packet_in reason is always correct.
And following this reasoning there seems to be no need to fix it up.

What there is a need for, which this patch does not address, is for packets
to be dropped rather than forwarded to OpenFlow1.3 controllers when a miss
occurs and an alternate behaviour has not been selected using Table Mod. I
plan to address this in subsequent patches.

As well as not been necessary, I believe that the logic that this patch
removes has not been executed because pin->generated_by_table_miss is never
true without a separate patch that I have proposed, "ofproto-dpip: Set
generated_by_table_miss for miss rule".  So I do not expect there to be any
run-time affect of this change.

This patch does not remove pin->generated_by_table_miss, although it
is now set but never used, as I plan to use when implementing the
OpenFlow1.3 drop-on-miss behaviour described above.

Cc: YAMAMOTO Takashi <yamamoto@valinux.co.jp>
Signed-off-by: Simon Horman <horms@verge.net.au>