datapath: Add phy_in_port to struct sw_flow_key
Add phy_in_port to struct sw_flow_key.

It is set it to the port number of the physical port that a packet was
received on.  This conveniently ignores the question of how to determine
the in_phy_port for cases such as tunnels where it may differ from in_port
but likely not be a port present in the bridge.

Compile tested only

Signed-off-by: Simon Horman <horms@verge.net.au>