ofp-util: Add in_phy_port to odp_flow_key_from_flow__()
Uses the same value for in_phy_port and in_port.

This conveniently ignores the question of how to determine
the in_phy_port for cases such as tunnels where it may differ
from in_port but likely not be a port present in the bridge.

Also update the comment above odp_flow_key_from_flow() and tests
to reflect this change.

Signed-off-by: Simon Horman <horms@verge.net.au>